+++ domain/solver/builders/__init__.py
@@ -1,4 +1,4 @@
 """L'import registra i builder nel BUILDERS."""
-from . import (grid, occupation, subject_buckets, subject_order,  # noqa: F401
-               subject_parts, time_counting, time_presence, time_sites,
-               unavailability, weight)
+from . import (grid, occupation, room_pool, subject_buckets,  # noqa: F401
+               subject_order, subject_parts, time_counting, time_presence,
+               time_sites, unavailability, weight)